Extract rest of struct helpes into pkg/structs
