+++ pkg/structs/structs.go
@@ -1,7 +1,12 @@
 package structs
 
 import (
+	"fmt"
 	"reflect"
+
+	"github.com/spf13/viper"
+
+	"github.com/radiohive/zimt/pkg/strings"
 )
 
 // Field represents single field of struct
@@ -42,4 +47,57 @@ func ExtractFields(arg interface{}) (fields []Field) {
 	}
 
 	return fields
+}
+
+// UnmarshalViper initializes config struct with viper values,
+// the struct fields are expected to have `viper` tags
+func UnmarshalViper(arg interface{}) {
+	fields := ExtractFields(arg)
+
+	if len(fields) == 0 {
+		return
+	}
+
+	for _, f := range fields {
+		t := f.Tag("viper")
+		if !f.Value.CanSet() || t == "" {
+			continue
+		}
+		switch f.Value.Kind() {
+		case reflect.Int:
+			f.Value.SetInt(viper.GetInt64(t))
+		case reflect.String:
+			f.Value.SetString(viper.GetString(t))
+		case reflect.Bool:
+			f.Value.SetBool(viper.GetBool(t))
+		}
+	}
+}
+
+// Report prints the config fields and values to the standard output
+func Report(arg interface{}, tag string) {
+	fields := ExtractFields(arg)
+
+	if len(fields) == 0 {
+		return
+	}
+
+	for _, f := range fields {
+		t := f.Tag(tag)
+		if t == "" {
+			continue
+		}
+		switch f.Value.Kind() {
+		case reflect.Int:
+			fmt.Printf("%s=%d\n", t, f.Value.Int())
+		case reflect.String:
+			if f.Tag("print") == "mask" {
+				fmt.Printf("%s=%q\n", t, strings.Mask(f.Value.String()))
+			} else {
+				fmt.Printf("%s=%q\n", t, f.Value.String())
+			}
+		case reflect.Bool:
+			fmt.Printf("%s=%t\n", t, f.Value.Bool())
+		}
+	}
 }
